src/server/NLUService/requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
blis==0.7.11
boto3==1.35.17
botocore==1.35.17
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.8
importlib_metadata==7.2.1
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
narwhals==1.6.2
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
preshed==3.0.9
protobuf==4.25.4
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl#sha256=a3a8eed42c600af17c70ed3c8720ffe2879b5353055e66fe0c998306329052fd
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
s3transfer==0.10.2
scikit-learn==1.3.2
scipy==1.13.1
setuptools==74.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.31.1
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validators==0.34.0
wasabi==1.1.3
watchdog==5.0.2
weasel==0.4.1
wrapt==1.16.0
zipp==3.20.1
